Add files via upload
First pass at code to recursively look down package hierarchies and display modules, classes, functions, builtins, and none-of-the-aboves.  Some Python documentation is good, some cryptic or nonexistent.  The goal is to see what packages offer and get a 'lay of the land' before trying to figure out how to do something